+++ libs/cardano-ledger-core/test/Main.hs
@@ -4,6 +4,7 @@ import qualified Test.Cardano.Ledger.AddressSpec as AddressSpec
 import qualified Test.Cardano.Ledger.BaseTypesSpec as BaseTypesSpec
 import qualified Test.Cardano.Ledger.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Common
+import qualified Test.Cardano.Ledger.Core.Tools as ToolsSpec
 import qualified Test.Cardano.Ledger.JsonSpec as JsonSpec
 import qualified Test.Cardano.Ledger.PlutusSpec as PlutusSpec
 import qualified Test.Cardano.Ledger.UMapSpec as UMapSpec
@@ -18,3 +19,4 @@ main =
       JsonSpec.spec
       UMapSpec.spec
       PlutusSpec.spec
+      ToolsSpec.spec